fix: move target_head_sha tests to separate file (750-LOC ceiling)
Extracts the 3 status-projection tests from watch.rs into
watch_status_tests.rs (test-named → file_size_invariant-exempt).
Brings watch.rs from 850 to 720 LOC (under the 750 ceiling).


Agend-Agent: claude-aef7c0
Agend-Task: t-20260714233249209912-64650-28
Agend-Branch: fix/ci-watch-target-head-status
Agend-Issued-At: 2026-07-14T23:32:52.728819+00:00
